requirements.txt
fastmcp>=0.2.0
mcp[cli]>=1.0.0
PyMuPDF>=1.24.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
openai>=1.40.0
numpy>=1.26.0
